+++ CHANGELOG.md
@@ -23,7 +23,7 @@ This plugin has functionality that is common to both ReSharper and Rider. It als
 
 - Fix incorrect redundant `SerializeField` attribute warning for property backing field ([#1016](https://github.com/JetBrains/resharper-unity/issues/1016), [#1464](https://github.com/JetBrains/resharper-unity/pull/1464))
 - Rider: Fix tooltip display for packages in Unity Explorer ([#1506](https://github.com/JetBrains/resharper-unity/pull/1506))
-
+- Unity Unit tests exploration. Easy to run all tests from solution. ([#1500](https://github.com/JetBrains/resharper-unity/pull/1500))
 
 
 ## 2019.3.2